Integrate ROS2 parser and template generator
Implementing nested type inclusion
